src: Reorganize files
